Extract hook loadout
